create nuget package
also update namespace to match nuget package metadata
