+++ Element.CloudDistributedLock.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33530.505
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Element.CloudDistributedLock", "Element.CloudDistributedLock\Element.CloudDistributedLock.csproj", "{21B603DE-A40B-4D78-AF46-527DC51D73C3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleApp", "ExampleApp\ExampleApp.csproj", "{3FCC734E-06A3-401D-989A-7BB3F06748C1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{21B603DE-A40B-4D78-AF46-527DC51D73C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{21B603DE-A40B-4D78-AF46-527DC51D73C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{21B603DE-A40B-4D78-AF46-527DC51D73C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{21B603DE-A40B-4D78-AF46-527DC51D73C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3FCC734E-06A3-401D-989A-7BB3F06748C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FCC734E-06A3-401D-989A-7BB3F06748C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FCC734E-06A3-401D-989A-7BB3F06748C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FCC734E-06A3-401D-989A-7BB3F06748C1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2911CA5A-29CE-4CE8-91EA-6F4C93E4F74A}
+	EndGlobalSection
+EndGlobal
