+++ Element.CloudDistributedLock/LockRecord.cs
@@ -1,0 +1,12 @@
+﻿namespace Element.CloudDistributedLock
+{
+    public class LockRecord
+    {
+        public string? id { get; set; }
+        public string? name { get; set; }
+        public string? providerName { get; set; }
+        public DateTimeOffset lockObtainedAt { get; set; } = DateTimeOffset.UtcNow;
+        public DateTimeOffset lockLastRenewedAt { get; set; } = DateTimeOffset.UtcNow;
+        public int _ttl { get; set; }
+    }
+}
