+++ Element.CloudDistributedLock/ServiceCollectionExtensions.cs
@@ -1,0 +1,40 @@
+﻿using Microsoft.Azure.Cosmos;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Element.CloudDistributedLock
+{
+    public static class ServiceCollectionExtensions
+    {
+        public static IServiceCollection AddCloudDistributedLock(this IServiceCollection services, string cosmosEndpoint, string cosmosKey, string databaseName, int ttl, Action<CloudDistributedLockProviderOptions>? configureOptions = null)
+        {
+            var cosmosClient = new CosmosClient(cosmosEndpoint, cosmosKey);
+            return services.AddCloudDistributedLock(CloudDistributedLockProviderFactory.DefaultName, cosmosClient, databaseName, ttl, configureOptions);
+        }
+
+        public static IServiceCollection AddCloudDistributedLock(this IServiceCollection services, CosmosClient cosmosClient, string databaseName, int ttl, Action<CloudDistributedLockProviderOptions>? configureOptions = null)
+        {
+            return services.AddCloudDistributedLock(CloudDistributedLockProviderFactory.DefaultName, cosmosClient, databaseName, ttl, configureOptions);
+        }
+
+        public static IServiceCollection AddCloudDistributedLock(this IServiceCollection services, string name, string cosmosEndpoint, string cosmosKey, string databaseName, int ttl, Action<CloudDistributedLockProviderOptions>? configureOptions = null)
+        {
+            var cosmosClient = new CosmosClient(cosmosEndpoint, cosmosKey);
+            return services.AddCloudDistributedLock(name, cosmosClient, databaseName, ttl, configureOptions);
+        }
+
+        public static IServiceCollection AddCloudDistributedLock(this IServiceCollection services, string name, CosmosClient cosmosClient, string databaseName, int ttl, Action<CloudDistributedLockProviderOptions>? configureOptions = null)
+        {
+            services.AddOptions();
+            services.AddSingleton<ICloudDistributedLockProviderFactory, CloudDistributedLockProviderFactory>();
+            services.Configure<CloudDistributedLockProviderOptions>(name, o =>
+            {
+                o.ProviderName = name;
+                o.CosmosClient = cosmosClient;
+                o.DatabaseName = databaseName;
+                o.TTL = ttl;
+                configureOptions?.Invoke(o);
+            });
+            return services;
+        }
+    }
+}
