+++ ExampleApp/Functions.cs
@@ -1,4 +1,4 @@
-using CloudDistributedLock;
+using Element.CloudDistributedLock;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using System.Net;
